amelioration hsitorique, ctrl-c, position du curseur, corrections leaks

+++ srcs/init_shell.c
@@ -8,14 +8,16 @@
 #include <unistd.h>
 #include <sys/stat.h>
 
+
 static int	check_terminal(t_cfg *cfg, uint8_t tty)
 {
-	if (cfg && (cfg->interactive = isatty(tty)))
+	if (isatty(tty))
 	{
 		if ((tcgetattr(tty, &cfg->term_origin) == FALSE))
 			perror("[TERM ORIGIN] ERROR TCGETATTR"); ///perror
 		return (1);
 	}
+	ft_ex(EXFD);
 	return (0);
 }
 
@@ -31,12 +33,14 @@ void		init_shell(char **env, char **av, int ac)
 	uint8_t		shell_terminal;
 	pid_t		shell_pgid;
 	t_cfg		*shell;
-	struct stat	stat;
+//	struct stat	stat;
 
-	if (fstat((shell_terminal = ttyslot()), &stat) == -1)
-		ft_ex(EXFD);
+
+//	if (fstat((shell_terminal = ttyslot()), &stat) == -1)
+//		ft_ex(EXFD);
+	shell_terminal = STDIN_FILENO;
 	shell = init_cfg(env, av, ac);
-	if (check_terminal(shell, shell_terminal))
+	if (shell->interactive && check_terminal(shell, shell_terminal))
 	{
 		while (tcgetpgrp(shell_terminal) != (shell_pgid = getpgrp()))
 			kill(-shell_pgid, SIGTTIN);
